Scripts08/S3DataMigration/requirements.txt
# Snowflake connector
snowflake-connector-python>=3.6.0

# PostgreSQL adapter
psycopg2-binary>=2.9.9

# Environment variable management
python-dotenv>=1.0.0

# Pandas for efficient data handling
pandas>=2.2.0

# Retry logic
tenacity>=8.2.3

# JSON schema validation
jsonschema>=4.20.0

